backend/requirements.txt
# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
python-socketio>=5.11.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Vector Database
chromadb>=0.4.22
# pinecone-client se usa opcionalmente si se necesita

# AI & ML
openai>=1.10.0
anthropic>=0.8.1
# transformers y torch son opcionales para SD local
# torch>=2.2.0
# diffusers>=0.26.0
# accelerate>=0.27.0
sentence-transformers>=2.3.0

# Image Generation
pillow>=10.2.0
replicate>=0.23.0

# TTS (Text-to-Speech)
edge-tts>=6.1.9
gtts>=2.5.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0,<4.1.0

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0
tenacity>=8.2.3

# Logging & Monitoring
loguru>=0.7.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
